requirements.txt
# ============================================================================
# COPD 数字孪生肺项目 - 依赖库列表
# 安装命令: pip install -r requirements.txt
# ============================================================================

# ----- 核心科学计算 -----
numpy>=1.21.0
scipy>=1.7.0

# ----- 医学影像处理 -----
nibabel>=4.0.0                # NIfTI 文件读写
pydicom>=2.3.0                # DICOM 文件读写
SimpleITK>=2.2.0              # 医学图像处理
# TotalSegmentator>=1.5.0     # 自动分割 (需单独安装，依赖 nnUNet)
# antspyx>=0.3.0              # ANTsPy (需从源码或 conda 安装)

# ----- 深度学习框架 -----
torch>=1.12.0
torchvision>=0.13.0

# ----- 3D 可视化 -----
pyvista>=0.37.0
vtk>=9.1.0

# ----- 图像处理与评估 -----
scikit-image>=0.19.0          # SSIM, PSNR 等指标
opencv-python>=4.5.0          # 图像处理
Pillow>=9.0.0                 # 图像 I/O

# ----- 数据处理与工具 -----
pandas>=1.4.0
tqdm>=4.64.0                  # 进度条
PyYAML>=6.0                   # 配置文件解析
matplotlib>=3.5.0             # 绑图

# ----- 开发工具 -----
pytest>=7.0.0                 # 单元测试
pytest-cov>=3.0.0             # 测试覆盖率
black>=22.0.0                 # 代码格式化
flake8>=4.0.0                 # 代码检查

# ----- Jupyter Notebook -----
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.0.0

# ----- 日志 -----
loguru>=0.6.0                 # 可选的高级日志库

